+++ src/java/entities/Student.java
@@ -3,10 +3,13 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entitis;
+package entities;
 
 import java.io.Serializable;
 import java.util.Set;
+import static javax.persistence.CascadeType.ALL;
+import static javax.persistence.FetchType.EAGER;
+import javax.persistence.OneToMany;
 
 /**
  *
@@ -14,6 +17,7 @@ import java.util.Set;
  */
 public class Student extends User implements Serializable{
     private LevelType levelType;
+    @OneToMany(cascade=ALL,mappedBy="student",fetch=EAGER)
     private Set<Mark> marks;
     
     //Setters and Getters
