+++ src/java/entities/UserType.java
@@ -3,14 +3,13 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entitis;
+package entities;
 
 /**
  *
  * @author Nerea
  */
-public enum LevelType {
-    BEGGINER,
-    MEDIUM,
-    EXPERIENCED
+public enum UserType {
+    TEACHER,
+    STUDENT
 }
